fix: transaction flow break
issue was because I was executing "CREATE TABLE" operation in b/w the
transaction which causes AUTO-COMMIT to transaction. Thus, is
transaction rollback was fired after that it won't revert the
inserts/deletes made in tables, as after commit each db operation is
treated as a single entity as transaction was commited automatically
before their execution.